requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
pytest==7.3.2
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
